mediawiki.page.watch.ajax: Move into separate dir to improve source map
Source maps are now enabled on www.mediawiki.org, and one of the
things that stand out are virtual files that appear directly in the
resources/src/ directory.

As per T193826, modules with multiple files should have their own
directory, and that includes virtual files when it comes to source
maps.

Bug: T47514
Bug: T193826
Change-Id: I49a9cc755a1d78a1468cf918abaadd9ea8dc1776