mediawiki.language: Move related modules and virtual files to separate dir
Source maps are now enabled on www.mediawiki.org, and one of the
things that stands out, is

> /w/resources/src/mediawiki.language.config.js

.... appearing direclty in /w/resources/src/. While the files
that reference it are under src/mediawiki.language/, the module was
incorrectly registered without a base directory.

While at it, give the "months" and "names" modules their own directory
as well to accurately reflect how the modules are defined.

As per T193826, modules with multiple files should have their own
directory, and with source maps having arrived, that now includes
virtual files.

Bug: T47514
Bug: T193826
Change-Id: Icce6310fb5e95aecfdf447df6ab2290068176381