MultiHttpClient: Unset $this->cmh after closing it
This should make no difference most of the time (the object is being
destructed anyway), but *if* the MultiHttpClient is somehow still used
after its destructor runs, we shouldn’t continue to reuse the cURL Multi
Handle that we already closed – getCurlMulti() should instead create a
new one.

Declare $this->cmh nullable to make Phan happy.

Bug: T288624
Change-Id: If74e3674b46868eef3e99bd338a824a6e7fe33cd