mediawiki.language: Use ES6 String.repeat method
String.prototype.repeat is significantly faster. Almost > 90% faster
than the local implementation

Note: Does not handle 'size' of Infinity (positive) and will throw.
I just leave it unhandled because it's obviously not allowed and the
replaced method did not attempt to handle it either (it just enters
infinite loop, which is even worse).

Change-Id: Ie3d8850225bb9b456d1a2cd4aa9a5a80de2bb2c4