mw.Api: Introduce abort signals
Add a `signal` option to mw.Api#ajax and related methods, which allows
the HTTP request(s) to be cancelled. The interface is inspired by the
Fetch API, and uses the same AbortController/AbortSignal classes.
Cancelling a request this way rejects it with a DOMException rather
than a 'http' error code, similarly to native fetch() function.

For compatibility (some browsers we support may not have a working
AbortController; it's a weird situation, see comments in the code),
provide our own implementation at mw.Api.AbortController.

Add better documentation for the existing style of cancelling promises
too, documented as mw.Api~AbortablePromise. It's not an actual class,
since the 'abort' method is added ad-hoc, just an alias for the docs.

mw.Api extensions can support both AbortSignal and AbortablePromise,
but it's a bit tricky to do. Add mw.Api.makeAbortablePromise()
to help with that, and explain how to do it in its documentation.

Update a few uses of the previous style, which either chained promises
in a complex way, or complained about the 'abort' method in comments.
I've left the simple cases unchanged, to show it's still supported.

(By the way, everything is called "abortable"/"aborted" rather than
"cancellable"/"cancelled", even though that's a nicer term, to avoid
confusion about whether these words have one or two 'l's.)

Depends-On: I7ff705af79760e33200a2dfbfdd05138c18bb110
Needed-By: I3a3ca1b32af58f396b66802a3d435b4c9d137e14
Needed-By: Ia803b4eab766768c2c8a096c308958b48eb34af2
Bug: T346984
Change-Id: Iec338e9f595b452c19ce8e74eb81339fbce11640