Add Block::toArray()
There are 7 callers of CompositeBlock::getOriginalBlocks() in code
search, and all are following a pattern like

  $blocks = $block instanceof CompositeBlock
    ? $block->getOriginalBlock() : [ $block ]

or can be reorganised to follow such a pattern, which suggests the need
for a public interface along these lines.

So, add Block::toArray() which returns the list of blocks in a
CompositeBlock, or wraps the single block in an array if the block is
not a CompositeBlock.

Also, remove the need for a phan-var override by moving the instanceof
DatabaseBlock check from shouldTrackBlockWithCookie() to its caller.

Bug: T345683
Change-Id: If769db9a831f63a79b57fb3bc9c7cece511c70c3