config: replace unused Logger in EtcdConfig with native trigger_error
Introduced in 9b459d29e0 (I60914d31c2, T156924) and
later extended in cde44feba3 (Ic5bdcb0cae).

There is no details about the fatal error in Logstash, because
EtcdConfig is by design a standalone class for use prior to any
service wiring, site configuration, or logger Spi can be used.

That's why EtcdConfig::setLogger() is unused and thus even the "error"
level messages go nowhere (NullLogger).

Rewrite these messages to instead surface via the native PHP error
log, the same way that our first-party and third-party standalone
libraries do, and as we do with deprecation warnings already.

Given that there isn't a way to inject this into MultiHttpClient,
we lose that visibility. However, not really because again setLogger
was never called in practice. And, the only error message
MultiHttpClient logs is the HTTP request failure, which is already
returned by $this->http->run() as $rerr to EtcdConfig and logged
by this patch.

Bug: T346971
Change-Id: I227ead395513877da88dd8b8abbd5ba7be25d5c7