Prevent empty namespace list in Special:Random
If none of the comma-separated subpages are valid namespaces, keep using
the default namespaces instead of passing an empty list into the rdbms
layer (where it causes an exception).

Also fix the coverage of the test, which got broken by the namespacing
in change I735222c437 (commit ea2979b8ca).

Bug: T347078
Change-Id: I461156f4150a1c5dd9824d27d51c58b10c22d3f5