phan: Re-enable class_alias support
This was disabled in I51dca12be9c77049f77563d9bf0edd07928c2300. The
rationale is still valid; however, without class_alias support we cannot
add tests for taint-check to verify that it knows about both the old and
the new class names. The benefits of being able to test taint-check
would currently outweigh the benefits of not using aliased classes.

Set enable_class_alias_support to true again, but add a TODO comment
that this should ideally be switched back to false if we can do that
without breaking taint-check tests.

Bug: T321806
Change-Id: I4177c7578cd84ce28d4ca359f73f89b6899bda68