Add taint-check annotations to database methods
These were copied verbatim from MediaWikiSecurityCheckPlugin.php.

The exception is selectRowCount(), whose return value was hardcoded as
tainted in taint-check, but that's clearly not the case because it
returns just an integer.

Suppress a new issue in DbTestPreviewer that wasn't previously spotted
because the code ends up using IReadableDatabase::selectRow, but the
hardcoded data in taint-check only considers IDatabase and concrete
classes, not IReadableDatabase.

Bug: T321806
Change-Id: I0c59d286c50de4ea79571242649bdda8aec67b57