Add another batch of taint-check annotations
Copied from MediaWikiSecurityCheckPlugin.php.

Duplicate annotations from Xml::encode* to the corresponding
Html::encode* methods, given that these were moved recently but not
hardcoded in taint-check.

As the only difference, remove the HTML taintedness type from the return
value of Message::rawParams. If the argument is unsafe, it's reported
immediately thanks to exec_html. Else, it does not contribute to the
taintedness of the return value.

Bug: T321806
Change-Id: I5ed340e1d127fb3eab6d6f9b905693d05a393360