MediaWiki: Use `$host` instead of trying to access array (again)
Above on line 1262, we already access the assoc array via the named
index 'host' to get the host name or default to null if we don't have
one.

So, we don't need to try doing the same thing again as we might hit
an error trying to access a key that doesn't exist meanwhile we've
already handled this above and defaulted to null. Just use the variable
in this case.

This fixes a phan warning which affected the patch on-top of this one.

Change-Id: I80c47017dc04946d3be0d0b0e0a73d0b62834155