ResourceLoader: Enable source maps by default
Disabled by default in WMF production, and currently enabled on
www.mediawiki.org and other group0 wikis.

To gain exposure through WMF CI and to provides its value immediately
to local development, enable it in MW core also.

If issues come up prior to 1.41.0, we can switch this back off,
and instead enable it in DevelopmentSettings.php to limit it to CI
and local dev.

Bug: T47514
Change-Id: I0cb58631b465bbd32ce70f9daae039d420d6bb3d