DerivedPageDataUpdater: Only do RefreshLinks on file page redirect changes
Previously, we scheduled recursive RefreshLinks jobs on every file
page edit, but it's only necessary when the page's redirect target has
changed. Optimize this, and only schedule them when the page is or was
a redirect.

(We still do it unnecessarily when a redirect page is edited without
changing the target, but that's more difficult to detect at this
point.)

Bug: T346383
Change-Id: Ie629881c94a7a81ae7bf597aab4587921d1c25ae