Annotate FormOptions::getValue as returning an unsafe value
Taint-check can actually determine this on its own, but this way the
error messages are more concise because they don't go into the details
of why the return value is tainted.

Note that the return value may actually be safe depending on its type,
but that's not easy for taint-check to know, and a core principle of
taint-check is to prefer false positives to false negatives.

Change-Id: I1e690f535a144cc53cbf3483a8f1d7d8d8eb4519