Clarify in inline docs how redirect URLs are expanded
Also fix remark about standards. RFC 2616 required an absolute
URL for the Location: header but later versions staring with
RFC 7231 do not, including RFC 9110 [1] which is the version
that made it to Internet Standard status; the WHATWG Fetch
spec [2] also agrees.

Nevertheless, there is no harm in expanding redirect URLs, and
maybe improves interoperability with old clients, so don't
change the logic, just adjust the comment.

[1] https://datatracker.ietf.org/doc/html/rfc9110#name-location
[2] https://fetch.spec.whatwg.org/#concept-response-location-url

Change-Id: Ifcc2e8ae3912fa5f704ede23857e66054b1b49e4