DifferenceEngine: Separate previous/next from diff itself
Having these controls in the diff makes it very difficult to
adapt the diff page for mobile screens. It is proposed we add
links at the top of the diff page, that are initially hidden.

These will initially be used in Minerva skin and possibly
upstreamed at a later date for all skins.

Change:
* Add previous / next breadcrumbs at top of diff page

Bug: T347780
Change-Id: Iae6c140cde54d807204e6d2b8c7b6599f75df89e