StatusValue: Improve logging/debug output with multibyte characters
Use mb_str_split and mb_substr to improve the chances that the fancy
table lines will line up, and to avoid corrupting multibyte characters
at boundaries. Remove max lengths in the sprintf format.

I'm not sure if this is the ideal output format, but I don't have the
heart to replace it.

Change-Id: Iff9ba98a12490938fcfefdeaebd5f1d82cd0ce7b