rdbms: Replace custom encoder/decoder with native strtr()
While this is not really a bottleneck the code path is hot enough to
show up in some of the profile runs I did. This is called tens of
thousands of times. This is notable because of the expensive, nested
loops as well as because the function call overhead in PHP is quite
big. Turns out we can replace all this with a native function that's
made for exactly this: searching for tokens and replacing them.

This code is well covered with tests.

Change-Id: I3e9fbbb162b941ecb14903dd7bde8f46bb071284