MagicWord: Make use of native count feature in preg_replace
The only reason for the callback was to check if a replacement
happened. We can do the same with the native $count feature.

Also inline a trivial sort() callback. It was effectively @internal
as it was obviously never meant to be used from outside of the class.

Both methods have been marked as public in I3588d9d. I think this
was just a mistake.

Change-Id: Ie33d4347f710fe227f9cbd8833c192c11d6e7c89