debug: Optimize createLogger() and set UTC timezone by default
Remove the need for wmf-config to perform dynamic 'calls'. It
currently uses this to make createLogger() call setTimezone().

The reasons for this are not WMF-specific, and thus can be done here.
In addition, Monolog\Logger supports a timezome argument so that we
don't need to construct two objects for every logger (one in
Monolog\Logger::__construct, and then another to replace it when
calling setTimezone), and also remove the setTimezone call and
skip the dynamic call for it.

The same call overhead can also be removed for the processors
and handlers array. Instead of calling push on each item, we can
pass these to the constructor directly. This does mean the order
will now be reversed but this is imho less surprising than before
(i.e. call from first to last, instead of reversed), and to my
knowledge none of our processors or handlers depend on side-effect
from previous ones.

Bug: T99581
Bug: T116550
Change-Id: Ib300c01c886dc8916413db65078a8356fd40a5c1