filebackend: Allow uploading files up to 32 GB with FSFileBackend
The SwiftBackend theoretical max file size was increased to 5 GB

Sysadmins of MediaWiki are in a better position to judge what
is acceptable performance-wise for uploaded files than we are. The
answer presumably depends significantly on the hardware mediawiki
is running under. This increases the max limit to 32 GB (previously 4GB).
With files this big, operations such as copying and calculating SHA1
do become expensive. However everything still works, albeit with
some delay (I tested on a laptop with HDD, I imagine things are
significantly faster with an NVMe drive).

This changes the overall FileBackendStore limit which affects all
subclasses that haven't overriden the limit, including those from
extensions. Swift was set to 5GB as that backend has to change the
way it uses the Swift API for objects larger than 5GB.

This should not affect ordinary users as $wgMaxUploadSize is by
default only 100 MB in MediaWiki (4GB in Wikimedia).

Bug: T191805
Change-Id: Id091d51b620edbeea2d9df8c75422807bfc7aea0