parser: Improve PHPDoc type hints in MagicWord… classes
Intentionally split across multiple patches. This is only about
documentation and impossible to break anything (other than Phan).

MagicWordArray::matchAndRemove is particularly confusing because the
documentation and structure of the returned array make it look like
it would support parameters. But it never (!) did.

The method was added like this in 2008 via commit 269a9103 (r31113).

There was always only a single caller in the Parser class. The
parser never used the array values, only the keys (via isset). Which
makes sense because that code in the parser is about "double
underscore" magic words (e.g. __NOTOC__). These don't support
parameters anyway.

Change-Id: Ife92fc3d6d5b03606ba2b209a886cadef3451fea