Don't cache trivial strings in the MagicWord class
4 strings are currently cached. But 3 are trivial derivatives from
the base one. This makes the code look more complicated than it is.
And we are wasting memory without knowing if the other strings are
ever needed.

These methods are not "hot", i.e. they are not called in a loop or
so.

Note the "initRegex" method removed in this patch was marked as
@internal. This means we really don't need to mention it in the
release notes. It makes zero sense to call it from the outside anyway.

Change-Id: I8606daaebc0dd5a8c7da5333841599c80763d75d