ApiSandbox: Only use POST and multipart/form-data when necessary
While it should always be valid to send any request as POST and
multipart/form-data, even those that are ordinarily sent as GET or
application/x-www-form-urlencoded, there are sometimes subtle
differences in behavior. For example, on Wikimedia wikis, POST
requests are always routed to the main datacenter, while GET requests
are not. ApiSandbox behaving like this is confusing when you're trying
to debug multi-DC bugs.

Therefore, only use POST and multipart/form-data when necessary.
API paraminfo already tells us when we need to do that.

Also display a note below the copyable parameters when these
mechanisms have to be used (and have been used) to send the request.

Change-Id: I5ed643dd2af8774c4205a3777cd8747ea29bbb23