SearchSuggestionSet: update pageMap when shrinking
the pageMap is a cache of page ids often used to quickly verify the
presence of a page in the set.
When shrinking this map should be updated to properly reflect the set of
page ids.
Not keeping pageMap updated caused a bug to appear when the exact match
is found by the underlying search engine at position limit+1 (overfetch
technique). Shrinking the set to limit did not properly remove the exact
match entry from pageMap which prevented the prepend method from adding
the exact match as part of the SearchExactMatchRescorer.

Bug: T347786
Change-Id: Id524cc5af8e62e64f0a680b8e08078d14872f97d