specials: Use the new page name for comment formatting
When a new talk page is created with section=new there is a section name
in the comment message linking to the section.
After a move of the new page it links to the old page while the section
is on the new page, this avoids a redirect (or a broken link in case of
suppress redirect).

Bug: T348665
Change-Id: Ic5ba84115a8a032c615116ca3ffb0ceeb0589151