mw.hook: Register handlers before running from memory
Most of the time we can expect a handler to be registered in
`fns` by the time it runs, which means handlers can use
mw.hook.remove to unregister themselves. This is useful if
you only want a handler to run once.

However when a hook has already been fired before the
handler is added, the hook is fired from `memory` immediately,
before we've even stored it in `fns`, so calling mw.hook.remove
doesn't work.

Change-Id: I545354e8c9a48756332e2d9a673a22dc5b0c16b8