Edit Recovery: Don't load when undoing edits
Don't load Edit Recovery when undoing an edit, because the
editor will only ever want to see the actual page content.

Also move the edit-conflict check from the client-side to PHP,
so that the checking for places that Edit Recovery shouldn't be
used is kept together in one place (and avoid unnecessarily
loading the JS).

Bug: T347869
Change-Id: Icd8cc8b5d221969595c268c8c0c3ff59aef4dc80