filebackend: Turn protected properties into actual constants
This is a direct follow up to Id0e4b0d from 2019 where these
"constants" have been introduced. I believe the reason for using
protected properties was that we still had to support PHP 7.0 back
then, but class constants can only be marked as protected since
PHP 7.1.

According to codesearch these are apparently meant to be internal
to the filebackend classes in core and not used anywhere else.

Change-Id: I55683e3540f8d76324370efecb98d08e07547c28