filebackend: Fix and add missing @return documentation
array[]|bool[]|null[] means: It's either an array of arrays, an array
that contains only bools, or an array where all values are null. But
this is not what's happening here, as far as I can tell.

This minor fixup is motivated by T348688, but doesn't solve it.
However, now it's more obvious that the values that trigger the
reported error are "null", i.e. some instance of self::$RES_ERROR.

Bug: T348688
Change-Id: Ia9debe97d21a3abff7bff75f16a94bec8be6ddc5