ParserOutputAccess: Fix local cache when page is edited within the process
When the latest revision of a page was requested from
ParserOutputAccess, then the page was edited within the same
PHP process, and then the new latest revision of a page was
requested, it would return the result for the old latest
revision. The $isOld checks did not prevent this, because at
the time of checking, the two revisions were both the latest.

Compare the latest revision ID (not just whether it's the
latest revision) to avoid this.

Bug: T349033
Change-Id: I386b4a9b791065fb39dcfb2cb6df9f321d540ae1