Edit Recovery: only save when form contents has changed
Rather than only tracking the change events of the form fields as
a means of determining whether the form has changed, monitor the
actual form fields' contents. This change saves the initial form
state into a variable to which it is then compared on each change
event, and if it differs it's saved but if it is unchanged then
the saved data is deleted.

Bug: T343089
Change-Id: Ic639b69bf62ed83d16d6a68ca1cc870607335243