Edit Recovery: discard section edit correctly
When deleting data in the abandonedit handler, we were passing the
section name to storage.deleteData(), but it wasn't being used.

This fixes that, and also adds the section name to the criteria
for deleting data in storage.deleteExpiredData(), meaning that
section or whole-page data isn't deleted when other records for
the same page are deleted.

Bug: T344410
Change-Id: I1178ba28c0d18e2e3728902d35c3145f3b26a7b9