Replace deprecated Codex legacy token
Replacing `@color-base--subtle` with `@color-subtle`.
Note, that `color-subtle` is using slightly darker `color.gray600` now
with design sanctus. It's a very _subtle_ difference :)
Note also, that this is the first core change for getting rid of
deprecated legacy tokens, the breakpoint change will need more
thorough testing and will be done in a separate patch.

Bug: T331403
Change-Id: I15c30269eb53a4efe7bd1d49877b4fef2dfadd6a