specials: Remove calls to UserLoginComplete hook with `direct=false`
When `direct` is set to false, it means we don't want to do anything
immediately but instead just redirect to the "return to" page or the
main page because the user is already logged in.

So it only really makes sense to call this hook in places where direct
is set to true (on real successful login) as this happens only once per
user login.

This patch moves consumers of the UserLoginComplete hook to only places
where there is a guarantee that `$direct` is set to true.

Bug: T349791
Change-Id: I8a2c91e1f173cf7b36f3dfec169e1d118277f057