Remove check for irrelevant setting: wgDisableCookieCheck
In https://phabricator.wikimedia.org/rMW3617c982c9db793515818e1468fa827ae5880358,
the "cookie redirect check" was removed but the setting that controlled
this set was not removed alongside.

Also, double checking that this setting is not configured in production
via our prod configs and always defaults to false is the case today.

Bug: T349791
Change-Id: Ia2a57b57c4c0a9848798f469181738646e0f8dc8