mail: Link to page log if not to a diff
When a page has been marked for translation or marked as reviewed, the
parameters that enotif receives include pageStatus='change', which is
the same as for revisions of a page's content. Thus the same wording in
the mail's body is used. This resulted in confusion for users ("Why did
I receive an email notification when there is no change in the page
revision history?").

This change aims to avoid that confusion by linking to the page log when
there is no revision to link to. This will at least happen for deletions
and for special changes such as those I mentioned earlier.

Bug: T115183
Bug: T145933
Change-Id: I12c09759f87f32f0df94d8441021b16e0e10c91c