Rewrite MagicWordArray::matchAndRemove to use single preg_… call
… instead of running the same (possibly expensive) regular expression
twice. The new code does the same as before, as proven by the tests.

I leave the error handling untouched for the moment, even if we know
from T321234 that the probably only reason the preg_… calls here can
fail is broken UTF-8.

The error message still mentions "preg_match_all". This is intentional
to make it easier to find old and new errors under the same name in
logstash.

I benchmarked this and it's indeed a bit faster. About 25% faster.
However, performance was not the main motivation for this patch but
the readability of the code.

Bug: T321234
Change-Id: I5b7a04abc008dd095dc87d0a05d06e061eb8d6b2